djangoProject2/library/views.py
from rest_framework.viewsets import ModelViewSet
from .serializers import ReaderSerializer, Reader, BookSerializer, Book, AuthorSerializer, Author


class AuthorViewSet(ModelViewSet):
    queryset = Author.objects.all()
    serializer_class = AuthorSerializer


class BookViewSet(ModelViewSet):
    queryset = Book.objects.all()
    serializer_class = BookSerializer


class ReaderViewSet(ModelViewSet):
    queryset = Reader.objects.all()
    serializer_class = ReaderSerializer

