pyproject.toml
[tool.poetry]
name = "djangoproject2"
version = "0.1.0"
description = ""
authors = ["evgeniatimoshkina <tea302@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
pillow = "^9.5.0"
envparse = "^0.2.0"
docker = "^6.1.3"
docker-compose = "^1.29.2"
psycopg = "^3.1.9"
django-extensions = "^3.2.3"
werkzeug = "^2.3.6"
django-phonenumber-field = "^7.1.0"
drf-spectacular = "^0.26.2"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
